move methods from posts to posts/data
